+++ java/demo/src/main/java/com/google/phonenumbers/demo/ResultsServlet.java
@@ -24,10 +24,10 @@ import static java.util.Locale.ENGLISH;
 import com.google.i18n.phonenumbers.NumberParseException;
 import com.google.i18n.phonenumbers.PhoneNumberUtil;
 import com.google.i18n.phonenumbers.Phonenumber.PhoneNumber;
-import com.google.phonenumbers.demo.helper.ErrorRenderer;
-import com.google.phonenumbers.demo.helper.ResultFileRenderer;
-import com.google.phonenumbers.demo.helper.ResultRenderer;
 import com.google.phonenumbers.demo.helper.WebHelper;
+import com.google.phonenumbers.demo.render.ErrorRenderer;
+import com.google.phonenumbers.demo.render.ResultFileRenderer;
+import com.google.phonenumbers.demo.render.ResultRenderer;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Locale;
